refactor: move examples in src/plothist and start testing each example script in pytest
