+++ src/plothist/examples/func_1d/1d_fct_stacked.py
