+++ src/plothist/examples/func_1d/README.rst
