+++ tests/test_examples.py
@@ -9,10 +9,10 @@ mpl_image_compare_kwargs = {
 
 
 @pytest.mark.mpl_image_compare(
-    filename="1d_fct.png",
+    filename="fct_1d.png",
     **mpl_image_compare_kwargs,
 )
 def test_plot_ttt():
-    from plothist.examples.fct_1d import make_figure
+    from plothist.examples.func_1d.fct_1d import make_figure
 
     return make_figure()
